requirements.txt
streamlit==1.38.0
fastapi==0.111.0
uvicorn[standard]==0.30.3
httpx==0.27.0
pyyaml==6.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pydantic==2.8.2
pypdf==4.3.1
pypdf2

requests~=2.32.5
python-dotenv~=1.1.1
openai~=2.2.0